Compatibility with Alfred This is a temporary fix really, I want to work on a package system that can use this main git repo as a base for adding these Alfred-specific workflow features, but this is a hotfix of sorts.
I went ahead and added the start of what will probably be an Odin-Alfred
library (Sigh), changed some functions (lazily) to support how Alfred
interprets data. This means I removed a lot of logging. This needs to be
remedied (Good thing I have a custom solution on the way!). Ignore the
ds store. I... am sorry?

There's A LOT of cruft at this point that needs to be pared down. This
should hopefully reduce the latency to under 200ms?